Remove X-Powered-By HTTP response header in Azure
Add applicationHost.xdt to publish to Azure to remove the X-Powered-By
HTTP response header that cannot be removed from the site itself when
using ASP.NET Core as described by

https://azure.microsoft.com/en-gb/documentation/articles/web-sites-transform-extend/
to resolve #1. This file has to be manually published to each app via
FTP as it sits outside the site's folder.