Publish web.config to Azure
It turns out that the real cause of #1 was the custom web.config file
not being included in publishing, so dotnet publish made its own, which
then didn't include the setting to disable X-Powered-By. Fix this by
including web.config in the files to publish. This makes
applicationHost.xdt obsolete, so it can be deleted.